Add unfinished script for duplicating Docker env.
The goal is to set up a separate directory with all the necessary
tools, each at the exact version level specified by the HCP folk in
order to run the latest additions (e.g. MSMAll). Some of the scripts in
the pipe call for older versions of FSL, as one example, but we're going
to go on the very reasonable assumption that the HCP people have tested
their container for correctness with these versions.